Move auth and auth-db into lib
