+++ src/lib/auth/index.ts
@@ -1,7 +1,7 @@
 import NextAuth from "next-auth";
 import GitHub from "next-auth/providers/github";
 import { MongoDBAdapter } from "@auth/mongodb-adapter";
-import client from "@/lib/db";
+import client from "./db";
 
 export const { handlers, auth, signIn, signOut } = NextAuth({
   basePath: "/auth",
